Úpravy v aplikacích ohledně vláken. Kompletně dodělanej výpis pingu. Předělanej WorkerThread, aby se nemoh budit, když už je vlákno mrtvý. Přidělaný metody na pojmenovávání vláken.

+++ psimulator2/src/applications/CiscoPingApplication.java
@@ -50,7 +50,7 @@ public class CiscoPingApplication extends PingApplication {
 	}
 
 	@Override
-	protected void handleIncommingPacket(IpPacket p, IcmpPacket packet, long delay) {
+	protected void handleIncommingPacket(IpPacket p, IcmpPacket packet, double delay) {
 
 
 		Logger.log(this, Logger.DEBUG, LoggingCategory.PING_APPLICATION, getName()+" handleIncommingPacket, type="+packet.type+", code="+packet.code+", seq="+packet.seq, packet);
